Refactoring and integrating react with api
