+++ demo/src/frontend-legacy/pages/display-plants.html
@@ -1,4 +1,3 @@
-
 <!DOCTYPE html>
 <html>
   <head>
@@ -11,18 +10,32 @@
     <div id="mydiv"></div>
 
     <script type="text/babel">
-      import React, { useState, useEffect } from 'react';
-      import ReactDOM from 'react-dom';
+      const { useState, useEffect } = React;
+      const { createRoot } = ReactDOM;
 
       function PlantsTable() {
           const [plants, setPlants] = useState([]);
+          const [error, setError] = useState(null);
       
           useEffect(() => {
               fetch('/api/plants')
                   .then(response => response.json())
-                  .then(data => setPlants(data))
-                  .catch(error => console.error('Error fetching plants:', error));
+                  .then(data => {
+                      if (Array.isArray(data.data)) {
+                          setPlants(data.data);
+                      } else {
+                          throw new Error('Data format is incorrect');
+                      }
+                  })
+                  .catch(error => {
+                      console.error('Error fetching plants:', error);
+                      setError(error.message);
+                  });
           }, []);
+      
+          if (error) {
+              return <div>Error: {error}</div>;
+          }
       
           return (
               <div>
@@ -63,7 +76,8 @@
           );
       }
 
-      ReactDOM.render(<PlantsTable />, document.getElementById('root'));
+      const root = createRoot(document.getElementById('mydiv'));
+      root.render(<PlantsTable />);
     </script>
 
   </body>
